Prevent deleting linked contacts

+++ resources/views/livewire/contacts/show.blade.php
@@ -14,6 +14,11 @@ new #[Layout('layouts.app')] class extends Component {
 
     public function delete()
     {
+        if (! $this->contact->canBeDeleted()) {
+            session()->flash('error', $this->contact->deletionBlockMessage());
+            return null;
+        }
+
         $this->contact->delete();
 
         session()->flash('status', 'Contato removido.');
@@ -26,6 +31,10 @@ new #[Layout('layouts.app')] class extends Component {
 <div class="flex flex-col gap-6">
     @if (session('status'))
         <x-jr.alert variant="success">{{ session('status') }}</x-jr.alert>
+    @endif
+
+    @if (session('error'))
+        <x-jr.alert variant="error">{{ session('error') }}</x-jr.alert>
     @endif
 
     @php
